+++ lib/govspeak/translation_helper.rb
@@ -1,5 +1,5 @@
 module Govspeak
-  module TranslationHelper
+  class TranslationHelper
     def self.t_with_fallback(key, **options)
       if I18n.available_locales.none?(options[:locale]&.to_sym)
         options[:locale] = I18n.default_locale
@@ -13,6 +13,16 @@ module Govspeak
       end
 
       I18n.t(key, **options)
+    end
+
+    def self.supported_locales
+      I18n.available_locales.map do |locale|
+        {
+          code: locale,
+          english_name: I18n.t("english_name", locale:),
+          native_name: I18n.t("native_name", locale:),
+        }
+      end
     end
   end
 end